requirements.txt
streamlit
python-dotenv
azure-storage-blob
azure-ai-formrecognizer